# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5f778259a3e049112322bb79
  title: cocKR
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  "/user/{id}/changePassword":
    post:
      x-skaffolder-id: 5f778259a3e049112322bb86
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5f778259a3e049112322bb83
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles:
        - 5f778259a3e049112322bb85
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
  /user:
    post:
      x-skaffolder-id: 5f778259a3e049112322bb89
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5f778259a3e049112322bb83
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5f778259a3e049112322bb8a
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5f778259a3e049112322bb83
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5f778259a3e049112322bb88
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f778259a3e049112322bb83
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f778259a3e049112322bb87
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f778259a3e049112322bb83
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5f778259a3e049112322bb8c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f778259a3e049112322bb83
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      x-skaffolder-id: 5f778259a3e049112322bb83
      x-skaffolder-id-db: 5f778259a3e049112322bb7a
      x-skaffolder-id-entity: 5f778259a3e049112322bb7b
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5f778259a3e049112322bb7e
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5f778259a3e049112322bb7f
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5f778259a3e049112322bb7d
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5f778259a3e049112322bb81
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5f778259a3e049112322bb80
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5f778259a3e049112322bb7c
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
    coc:
      x-skaffolder-id: 5f7811fba3e049112322bcf6
      x-skaffolder-id-db: 5f778259a3e049112322bb7a
      x-skaffolder-id-entity: 5f7811faa3e049112322bcf1
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        coc_id:
          type: integer
          x-skaffolder-id-attr: 5f7811faa3e049112322bcf5
          x-skaffolder-type: Integer
        first_name:
          type: string
          x-skaffolder-id-attr: 5f7811faa3e049112322bcf2
          x-skaffolder-type: String
        last_name:
          type: string
          x-skaffolder-id-attr: 5f7811faa3e049112322bcf3
          x-skaffolder-type: String
        shop_name:
          type: string
          x-skaffolder-id-attr: 5f7811faa3e049112322bcf4
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
  x-skaffolder-db:
    - x-skaffolder-id: 5f778259a3e049112322bb7a
      x-skaffolder-name: cocKR_db
  x-skaffolder-page:
    - x-skaffolder-id: 5f778259a3e049112322bb82
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
  x-skaffolder-roles: 
    - x-skaffolder-id: 5f778259a3e049112322bb85
      x-skaffolder-name: ADMIN